% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_data.R
\name{import_data}
\alias{import_data}
\title{Process NHS Sickness Absence Data}
\usage{
import_data(
  data_path = system.file("app/data_raw", package = "HospitalAbsences")
)
}
\arguments{
\item{data_path}{Character. The path to the folder containing the CSV files.}
}
\value{
A processed tibble with cleaned names, and narrowed scope
}
\description{
\emph{fct_import_data.R} reads all CSV files from the data structure,
cleans column names, narrows the scope of the data to only hospitals and community providers,
and creates columns appropriate for use in other modules.
}
